+++ src/react.rs
@@ -1,7 +1,7 @@
-//! The events that are intended to be read by the user to react to are stored here.
+//! The structs that are intended for the user too trigger or react too.
 //!
 
-use bevy_ecs::{component::Component, entity::Entity};
+use bevy_ecs::{component::Component, entity::Entity, event::Event};
 
 /// Marker to indicate this node is currently being highlighted by this tooltip
 /// When this component is added user should apply styling so it's obvious to the player
@@ -16,8 +16,28 @@ pub struct TooltipHighlighting {
     pub entity: Entity,
 }
 
-/// Marker to indicate that this `ToolTip` should not be despawned.
+/// Marker to indicate that this `Tooltip` should not be despawned.
 /// When this component is added user should apply styling so it's obvious to the player
 /// that the tooltip will not be despawned by timeout or pointer leaving.
 #[derive(Debug, Component)]
 pub struct TooltipLocked;
+
+/// Manually spawn a `Tooltip`, useful for icons users may click.
+/// Not currently supported for nested tooltips.
+#[derive(Event)]
+pub struct SpawnTooltip {
+    /// The term to lookup
+    pub term: String,
+
+    /// The entity spawning this, will quick return if existing tooltips
+    pub entity: Entity,
+}
+
+impl SpawnTooltip {
+    pub fn new(term: impl Into<String>, entity: Entity) -> Self {
+        Self {
+            term: term.into(),
+            entity,
+        }
+    }
+}